made the object save to local storage
