+++ products/index.html
@@ -13,9 +13,7 @@
 
 <body>
     <ul id='instrument'></ul>
-
-    
+  
     <script src='./products.js' type='module'></script>
-    <script src='../data-persistence/local-cart.js' type='module'></script>
 </body>
 </html>
