+++ products/renderfunction.js
@@ -1,3 +1,5 @@
+import findProducts from '../shopping-cart/find-by-id.js';
+
 const renderInstrument = (instrument) => {
     const li = document.createElement('li');
     li.classList = instrument.category;
@@ -35,8 +37,39 @@ const renderInstrument = (instrument) => {
 
     const button = document.createElement('button');
     button.textContent = 'Add to Cart';
-    li.appendChild(button);
+    button.value = instrument.code;
+    button.addEventListener('click', function() { 
 
+        let json = localStorage.getItem('CART');
+        let cart;
+        if (json) {
+            cart = JSON.parse(json);
+        }
+        else {
+            cart = [];
+        }
+
+        let lineItem = findProducts(cart, instrument.code);
+        
+        if (!lineItem) {
+
+            lineItem = {
+                code: instrument.code,
+                quantity: 1
+            };
+
+            cart.push(lineItem);
+        }
+        else {
+            lineItem.quantity++;
+            
+        }
+        json = JSON.stringify(cart);
+        localStorage.setItem('CART', json);
+    });
+
+    li.appendChild(button);
+    
     return li;
 };
